Rework the HTTP utilities for improved information preservation
Specifically, the 'request' function, which has been renamed to
'sendRequest', has two significant changes:

- It now returns the input request options along with the
  resulting IncomingMessage.
- The request body input now has to be wrapped in a Future to facilitate
  the former change: In order to return pure request options, the body
  must not yet have been initialized.

The reason for this rework is to faciliate the creation of new functions
that require *both* the Request and the IncomingMessage to perform their
tasks. Specifically, I am thinking of following redirects, where the
original request has to be re-sent with an altered URL.

To facilitate these changes, this commit introduces two new types:

- Request: The wrapper object for the request options.
- Response: The wrapper object for the Request/IncomingMessage pair.

Some functions have been reworked to operate on the Response type
instead of the IncomingMessage type.

Although this is a breaking change, for most uses of the library,
nothing will need to change in user land. This is because the interface
remained the same for most of the common use-cases, even though the data
types that flow through have been altered.