Add functions for following redirects
Adds a framework for redirection of requests. Add the core of the idea
is one new function: 'followRedirectsWith'. This function takes what I'm
calling a "rediction policy" in the form of a lambda that takes a
Response and produces a (new) Request from it ('Response -> Request').

Using the redirection policy and a max number, the followRedirectsWith
recursively follows the new Requests produced, until either the maximum
number of redirects is reached, or the policy produces a request which
has already been seen in the chain of redirections.

This last principle means we are protected against infinite redirect
loops, but it also gives plicies a way to "not redirect" by simply
returning the original request.

There are two built-in redirection policies:

- The 'defaultRedirectionPolicy' policy follows RFC2616 strictly.
- The 'aggresiveRedirectionPolicy' policy deviates a little from RFC2616
  to better facilitate fully automated programs.

These policies are composed from smaller policies using the matchStatus
function, and users have access to the underlying 'Response -> Request'
function to build their own policies, if they wish.